Split graph module, and several improvements to the graph datastructure
