+++ ir/src/ir2/graph/mod.rs
@@ -1,0 +1,7 @@
+mod graph;
+mod link;
+mod nodes;
+pub use graph::{Child, Graph, Leaf, Node, Parent};
+pub use link::{BackLink, Link};
+
+pub use nodes::{Add, Function, NodeTypes};
